+++ src/components/advanced-audio-desc-toggle/index.js
@@ -1,0 +1,1 @@
+export {AdvancedAudioDescToggle} from './advanced-audio-desc-toggle';
