+++ src/components/audio-menu/index.js
@@ -1,0 +1,1 @@
+export {AudioMenu} from './audio-menu';
