+++ src/components/captions-menu/index.js
@@ -1,0 +1,1 @@
+export {CaptionsMenu} from './captions-menu';
