+++ src/components/index.js
@@ -56,6 +56,11 @@ export {CopyButton} from './copy-button';
 export {ButtonControl} from './button-control';
 export {Button} from './button';
 export {ToggleSwitch} from './toggle-switch';
+export {AudioMenu} from './audio-menu';
+export {CaptionsMenu} from './captions-menu';
+export {SpeedMenu} from './speed-menu';
+export {QualityMenu, HeightResolution} from './quality-menu';
+export {AdvancedAudioDescToggle} from './advanced-audio-desc-toggle';
 
 export {PlayerArea, withPlayerPreset, Remove} from './player-area';
 export {VideoArea} from './video-area';
@@ -65,7 +70,6 @@ export {InteractiveArea} from './interactive-area';
 export {Keyboard as KeyboardControl} from './keyboard';
 export {Cast, Cast as CastControl};
 export {Fullscreen, Fullscreen as FullscreenControl};
-export {Language, Language as LanguageControl};
 export {PlayPause, PlayPause as PlayPauseControl};
 export {Rewind, Rewind as RewindControl};
 export {Forward, Forward as ForwardControl};
