--- src/components/language/index.js
@@ -1,1 +1,0 @@
-export {Language} from './language';
