+++ src/components/quality-menu/index.js
@@ -1,0 +1,1 @@
+export {QualityMenu, HeightResolution} from './quality-menu';
