+++ src/components/speed-menu/index.js
@@ -1,0 +1,1 @@
+export {SpeedMenu} from './speed-menu';
